requirements.txt
# Runtime dependencies
aiohttp==3.9.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2

# Google Drive API dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0